+++ RELEASE_NOTES.md
@@ -7,9 +7,10 @@ This version deprecates the `OnboardingPresenter` concept and adds a `tryPresent
 
 ### ✨ New features
 
-* `UrlTutorialPage` is a new tutorial page with a `URL`.
+* `TutorialPageInfo` is a new struct used by `TutorialPageTabView`.
 * `TutorialPageTabView` is a new view for presenting tutorial-based onboarding flows.
 * `TutorialPageTabViewStyle` can be used to style `TutorialPageTabView` views.
+* `UrlTutorialPage` is a new tutorial page with a `URL`.
 
 ### 💡 Behavior changes